+++ src/DAL/models/task.ts
@@ -5,7 +5,7 @@ import { ITaskGeometries } from './taskGeometry';
 
 export interface ITaskEntity {
   /** The auto-generated ID of the artifact. */
-  id?: number;
+  id: number;
   /** The catalog record id of the requested layer. */
   catalogRecordID: string;
   /** The unique job id of the task. */
@@ -23,21 +23,21 @@ export interface ITaskEntity {
   /** list of requested webhooks actions of the task. */
   webhook: Webhook[];
   /** Status of the task. */
-  status?: TaskStatus;
+  status: TaskStatus;
   /** The Description of the task. */
   description: string;
   /** Free keywords to describe the task by the user. */
   keywords: Record<string, unknown>;
   /** The estimated data size. */
-  estimatedDataSize?: number;
+  estimatedSize?: number;
   /** The estimated time for the task to be completed. */
   estimatedTime?: number;
-  /** The task current percentage. */
-  percentage?: number;
-  /** The reason that the task failed. */
-  reason?: string;
+  /** The task current progress. */
+  progress?: number;
+  /** The errorReason that the task failed. */
+  errorReason?: string;
   /** Automatically generated date when the given task was created. */
-  createdAt?: Date;
+  createdAt: Date;
   /** Automatically update the date when the any progress or status has changed. */
   updatedAt?: Date;
   /** Automatically generated date that the task will be expired . */